sql: define the tag table
An identifier, an optional parent tag and a name,
implementing a hierarchical tag system.
Very simple yet surprisingly powerful.

There are some uniqueness constraints to the tag attributes
but since nulls are involved I'll use partial indexes to express them.
It's not possible to use WHERE clauses in table CONSTRAINTs.