sql: create the metadata table
Connects data to tags in a many-to-many relationship.
Each row of this table represents an application of a tag to data.
There's also a value column which implements a key/value store
using the tag itself as the key. No type affinity is declared
and SQLite allows storing any type of value into the column. Neat!

There are some uniqueness constraints but they depend on whether a value
is present. Since nulls are involved, I'll use partial indexes for them.
It's not possible to use WHERE clauses in table CONSTRAINTs.