taggr: add default value of None to data argument
It's not actually necessary for the data itself to be in the database.
A data row can be just an anchor to attach tags to,
the data itself can be stored elsewhere.

Make it easy to insert new null data rows into the database
by not requiring a parameter to be passed.