sql: handle uniqueness constraint violations
Since there are uniqueness constraints on the tags and their relations,
conflicts will arise in case the same tag is inserted more than once.
The database will attempt to resolve the conflict in some way.
The default behavior is to abort the transaction and raise an error.

This maintains the integrity of the database by preventing duplicates.
Unfortunately, it makes the database needlessly hard to work with
since the user now has to worry about whether a record already exists
before attempting to insert it, complicating application logic.

For ease of use, insertions into the tags database should be idempotent:
attempts to insert data already present in the database should succeed
without producing errors of any kind and return the existing rows.

The first attempted solution was to replace the INSERT with an UPSERT
that simply ignores the error:

    ON CONFLICT DO NOTHING
    RETURNING id

The RETURNING clause was designed to provide easy access to the columns
that are filled in by the database itself such as the row ID.
It works perfectly when a new row is inserted in the database.
I hoped it would also work for the purpose of returning the row ID
of the existing record found by the UPSERT statement.

This hope proved to be a bit naive. Study of the documentation reveals:

> The RETURNING clause only returns rows that are directly modified
> by the DELETE, INSERT, or UPDATE statement.

Since I told the database to DO NOTHING in case of conflicts,
the RETURNING clause instead returned absolutely nothing.
This clearly wasn't going to work.

However, the documentation also contains a solution to this conundrum:

> A RETURNING clause for an UPSERT
> reports both inserted and updated rows.

The simplest way out is to update the record. Neat.
I was hoping to avoid additional I/O in case of duplicates
but it looks like that's not going to be possible for now.
I added the timestamp columns to at least get some value out of it.
This has the added benefit that it should also work on PostgreSQL,
easing a possible future migration or alternative backend.

In my search for alternative solutions,
I posted anonymously in the SQLite forums.
The thread converged to the aforementioned solution
but ended with a suggestion for a new SQLite feature
for explicitly supporting idempotent insertions.
Who knows, right?

I also asked in the SQLite and PostgreSQL IRC channels on libera.
People from the latter channel provided an interesting solution
involving a common table expression that selects the existing row
if the insertion result set is empty. Worked well in PostgreSQL
but it turned out SQLite does not support INSERT statements
in the body of a WITH statement. Unfortunately, I could not use it.

References:

    https://en.wikipedia.org/wiki/Idempotence
    https://sqlite.org/lang_conflict.html
    https://sqlite.org/lang_upsert.html
    https://sqlite.org/lang_returning.html
    https://sqlite.org/lang_with.html
    https://sqlite.org/forum/forumpost/ae74a8a1d0d1c94b