pyproject: add script entries
Now the package manager will generate scripts
that invoke the taggr command line interface.

I really want a simple and short "tag" command...
I searched the web but didn't find any other tool that dared to use it.
Is it really possible that I'll be the first?
I'll just go ahead and do it.