cli: submit file data for hashing in chunks
When reading data from files, the input stream is seekable
which enables calculation of the final blob size for SQLite.
This allows the code to insert the data into the database
in small chunks instead of reading it all at once into memory.
Each chunk is inserted into the database and submitted for hashing
at the same time and the results can be collected later.